Move binaryinfo and checksumutils to pkg
