+++ pkg/binaryinfo/BinaryInfo.go
