+++ pkg/gitutils/gitparameteroptions/GitRepositoryCreateTagOptions.go
@@ -3,8 +3,8 @@ package gitparameteroptions
 import (
 	"fmt"
 
-	"github.com/asciich/asciichgolangpublic/binaryinfo"
 	"github.com/asciich/asciichgolangpublic/logging"
+	"github.com/asciich/asciichgolangpublic/pkg/binaryinfo"
 	"github.com/asciich/asciichgolangpublic/tracederrors"
 )
